fix(types): export types correctly and reliably
