+++ src/index.ts
@@ -1,2 +1,3 @@
 export * from './chessGame';
 export * from './chessStudy';
+export * from './types';
